Fix advanced search not sorting or paging
`useEffect` must listen to `url` changes in order to respond to sorting and paging. Based on the investigation done to resolve this issue, `UseDirectory` is listening to the `url` here instead of `filters`. When the code was copied and re-implemented for the advanced search, it seems we forgot to include to listen for the `url` changes as well as for the `filters`. The code is already protected to only go and fetch data if the are filters applied only.